rename: 게시판 파일, 디렉토리 수정
