+++ matzipback/src/main/java/com/matzip/matzipback/board/command/application/controller/BoardCommandController.java
@@ -1,0 +1,4 @@
+package com.matzip.matzipback.board.command.application.controller;
+
+public class BoardCommandController {
+}
