+++ matzipback/src/main/java/com/matzip/matzipback/board/command/application/service/BoardCommandService.java
@@ -1,0 +1,4 @@
+package com.matzip.matzipback.board.command.application.service;
+
+public class BoardCommandService {
+}
