+++ matzipback/src/main/java/com/matzip/matzipback/board/command/domain/repository/BoardDomainRepository.java
@@ -1,0 +1,4 @@
+package com.matzip.matzipback.board.command.domain.repository;
+
+public class BoardDomainRepository {
+}
