+++ matzipback/src/main/java/com/matzip/matzipback/board/command/domain/service/BoardDomainService.java
@@ -1,0 +1,4 @@
+package com.matzip.matzipback.board.command.domain.service;
+
+public class BoardDomainService {
+}
