+++ matzipback/src/main/java/com/matzip/matzipback/board/command/infrastructure/repository/BoardInfraRepository.java
@@ -1,0 +1,4 @@
+package com.matzip.matzipback.board.command.infrastructure.repository;
+
+public interface BoardInfraRepository {
+}
