+++ matzipback/src/main/java/com/matzip/matzipback/board/command/infrastructure/service/BoardInfraService.java
@@ -1,0 +1,4 @@
+package com.matzip.matzipback.board.command.infrastructure.service;
+
+public class BoardInfraService {
+}
