+++ matzipback/src/main/java/com/matzip/matzipback/board/query/controller/BoardQueryController.java
@@ -1,0 +1,4 @@
+package com.matzip.matzipback.board.query.controller;
+
+public class BoardQueryController {
+}
