+++ matzipback/src/main/java/com/matzip/matzipback/board/query/mapper/BoardMapper.java
@@ -1,0 +1,4 @@
+package com.matzip.matzipback.board.query.mapper;
+
+public interface BoardMapper {
+}
