+++ matzipback/src/main/java/com/matzip/matzipback/board/query/service/BoardQueryService.java
@@ -1,0 +1,4 @@
+package com.matzip.matzipback.board.query.service;
+
+public class BoardQueryService {
+}
